Revert "Do not destroy consumer on cancellation"
